IN-421 refactor: add index import + moving logic MessagesMarkup into a hook
