+++ src/features/messenger/model/index.ts
@@ -1,0 +1,3 @@
+export { useMessagesMarkup } from './useMessagesMarkup'
+export { useDialogsList } from './useDialogsList'
+export { useSendMessageTextArea } from './useSendMessageTextArea'
