+++ src/features/messenger/ui/MessagesMarkup.tsx
@@ -1,35 +1,11 @@
-import { useEffect, useRef } from 'react'
-
-import { messagesApi } from '@/common/api/messenger.api'
+import { useMessagesMarkup } from '@/features/messenger/model'
 import { Message } from '@/features/messenger/ui/Message'
 import { ScrollArea } from '@byte-creators/ui-kit'
-import { useRouter } from 'next/router'
 
 import styles from './styles/Message.module.css'
 
 export const MessagesMarkup = () => {
-  const { query } = useRouter()
-  const dialoguePartnerId = Number(query.id) || 0
-  const { data, isLoading, isFetching } = messagesApi.useGetMessagesQuery({
-    dialoguePartnerId,
-  })
-  const [updateMessageStatus] = messagesApi.useUpdateStatusMessageMutation()
-  const { data: me } = messagesApi.useMeQuery()
-  const messagesEndRef = useRef<HTMLDivElement>(null)
-
-  useEffect(() => {
-    if (!isLoading && data && me) {
-      messagesEndRef.current?.scrollIntoView({ behavior: 'auto' })
-
-      const unreadMessages = data.items.filter(
-        m => m.receiverId === me.userId && m.status !== 'READ'
-      )
-
-      if (unreadMessages.length > 0) {
-        updateMessageStatus({ ids: unreadMessages.map(m => m.id) })
-      }
-    }
-  }, [data, isLoading, updateMessageStatus, me])
+  const { isFetching, isLoading, data, query, messagesEndRef } = useMessagesMarkup()
 
   if (isLoading || isFetching) {
     return null
