+++ src/features/messenger/ui/SendMessageTextArea.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from 'react'
 
-import { useSendMessageTextArea } from '@/features/messenger/model/useSendMessageTextArea'
+import { useSendMessageTextArea } from '@/features/messenger/model'
 import { Alert, Button, TextArea } from '@byte-creators/ui-kit'
 import { mergeRefs } from '@byte-creators/utils'
